NuSpec with a dependency on VS2013's Scaffolder
Closes #116

Scaffolding in 2013 is changing (and has changed) and won't be
supported. However, there is a vs2013 version of the T4Scaffolder and
this package depends on it.

Good news - This seems to work fine under VS3013 when the IDE is run as
Admin.
Bad news - Because this NuGet will have a dependency on a -pre
(prelease) package, then it becomes prerelease. That means there are TWO
packages. MvcMailer and MvcMailer-vs2013.

http://docs.nuget.org/docs/reference/versioning#Prerelease_Versions

That means, you would install it like this:

install-package MvcMailer-vs2013 -pre

Which in turn brings in the T4 pre dependency.

NOTE: I didn't mess with the rake file OR the publish. A NuPkg will get
make when you run rake as the work happens in CSProj (looks like the
rakefile was never completed to support NuGet?)

NOTE 2: I had to update NuGet.exe to latest.

TODO: You'll want to publish the *new* package as it's down thing (and
perhaps advertise that it exists by mentioning it in the EXISTING
version's NuGet package as well as in the Readme.md? I added that also,
if that's OK.